+++ lib/flutter_blue_plus.dart
@@ -14,6 +14,7 @@ part 'src/bluetooth_characteristic.dart';
 part 'src/bluetooth_descriptor.dart';
 part 'src/bluetooth_device.dart';
 part 'src/bluetooth_msgs.dart';
+part 'src/bluetooth_events.dart';
 part 'src/bluetooth_service.dart';
 part 'src/bluetooth_utils.dart';
 part 'src/flutter_blue_plus.dart';